requirements.txt
huggingface_hub==0.21.4
Jinja2==3.1.3
numpy==1.26.4
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
